avg_photometry.py : changes the name files are saved under so they can be properly globed later sort_photometry.py : changes made to assign_id function to make it compatible with make_CMD's needs, does not affect it's other functionality.
output locations in avg_photometry.py, load_photometry.py, and sort_photometry.py are now the full directory pathway.

make_CMD.py : Several parameters with defaults added to allow the way the results are displayed to the controlled by the user. Cluster membership was added using a membership list input by the user but if no list is given all results are plotted.